Fixed deprecation warnings in PHP 8+ (issue #30 and issue #32). (#33)
Thanks! Seems to work like a charm.
As I am no more using dokuwiki myself it is not easy for me to do a good test coverage and the set of tests I currently have are very basic... (should create a Dockerfile to be able to change easily PHP version and add some automated tests but always difficult to find the time...) so your contribution is very appreciated!